+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Pre-hooks stream output to stderr in real-time; post-hooks write to log files in `.jj/djo/logs/`
+
 ## [0.2.0] - 2026-04-09
 
 ### Fixed